scripts/data/processing/gcs/requirements_seg.txt
google-cloud-compute
google-cloud-storage
fire
tqdm
git+https://github.com/mattdeitke/moto3.git
boto3
numpy
pysrt
webvtt-py
openai-whisper
pydub